app/src/main/java/com/auttmme/moviecatalogue/ui/movies/MovieViewModel.kt
package com.auttmme.moviecatalogue.ui.movies

import androidx.lifecycle.ViewModel
import com.auttmme.moviecatalogue.data.MovieEntity
import com.auttmme.moviecatalogue.data.source.MovieCatalogueRepository
import com.auttmme.moviecatalogue.utils.DataDummy

class MovieViewModel(private val movieCatalogueRepository: MovieCatalogueRepository) : ViewModel() {

    private var movieId: Int = 0

    fun setSelectedMovie(movieId: Int) {
        this.movieId = movieId
    }

    fun getMovie(): MovieEntity {
        lateinit var movie: MovieEntity
        val movieEntities = DataDummy.generateDummyMovies()
        for (movieEntity in movieEntities) {
            if (movieEntity.movieId == movieId) {
                movie = movieEntity
            }
        }
        return movie
    }

    fun getAllMovies(): List<MovieEntity> = movieCatalogueRepository.getAllMovies()
}